i added functionality to A and D , i also added a title screen
